% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatz.R
\name{collatz}
\alias{collatz}
\title{Collatz Numbers}
\usage{
collatz(n, limit = 10000)
}
\arguments{
\item{n}{the initial integer}

\item{limit}{maximum number of members of the Collatz sequence to compute}
}
\value{
side effects
}
\description{
Find and graph the Collatz sequence from a given initial number.
}
\examples{
\dontrun{
collatz(1757)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
