% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courtSim.R
\name{courtSim}
\alias{courtSim}
\title{Appeals Court Simulation}
\usage{
courtSim(reps = 10000,
                seed = NULL,
                table = FALSE,
                probs = c(0.95, 0.94, 0.90, 0.90, 0.80))
}
\arguments{
\item{reps}{number of simulations to perform}

\item{seed}{The user may provide a seed-value for random-number
generation.}

\item{table}{Does the user want a table of the results?}

\item{probs}{Chance for each judge to make the right decision on any
given case.}
}
\value{
side effects
}
\description{
In this version all five judges vote independently.
}
\examples{
courtSim(seed = 3030)
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
