% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distExplore.R
\name{distExplore}
\alias{distExplore}
\title{Exploring Major Probability Distributions}
\usage{
distExplore(options = NULL)
}
\arguments{
\item{options}{Options that will be passed to \code{shiny::shinyApp}.}
}
\value{
side effects
}
\description{
Manipulate parameters of a chosen distribution.
}
\examples{
## distExplore()
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
