% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteBinom.R, R/drunkenSim.R
\name{drunkenSim}
\alias{drunkenSim}
\title{Drunken-Turtle Simulation}
\usage{
drunkenSim(steps = 1000, reps = 10000, close = 0.5,
                       seed = NULL, table = FALSE)

drunkenSim(steps = 1000, reps = 10000, close = 0.5,
                       seed = NULL, table = FALSE)
}
\arguments{
\item{steps}{the number of steps the turtle will take}

\item{reps}{number of simulations to perform}

\item{close}{the distance from the origin that counts as "close"}

\item{seed}{The user may provide a seed-value for random-number
generation.}

\item{table}{Does the user want a table of the results?}
}
\value{
side effects

side effects
}
\description{
A drunken turtle starts at the origin and takes unit steps,
turning through a randdom angle after each step.  We are interested in
the distribution of the number of close-returns to the origin, for a fixed
number of steps and a fixed measure of closeness.

A drunken turtle starts at the origin and takes unit steps,
turning through a randdom angle after each step.  We are interested in
the distribution of the number of close-returns to the origin, for a fixed
number of steps and a fixed measure of closeness.
}
\examples{
drunkenSim(seed = 3030)
drunkenSim(seed = 3030)
}
\author{
Homer White \email{homerhanumat@gmail.com}

Homer White \email{homerhanumat@gmail.com}
}
