% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drunkenSim2.R
\name{drunkenSim2}
\alias{drunkenSim2}
\title{Drunken-Turtle Simulation (Graph Version)}
\usage{
drunkenSim2(steps = 1000, seed = NULL)
}
\arguments{
\item{steps}{the number of steps the turtle will take}

\item{seed}{The user may provide a seed-value for random-number
generation.}
}
\value{
side effects
}
\description{
A drunken turtle starts at the origin and takes unit steps,
turning through a randdom angle after each step.  We graph the distance
from the origin as a function of step-number.
}
\examples{
\dontrun{
drunkenSim2(seed = 3030)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
