% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdExplore.R
\name{kdExplore}
\alias{kdExplore}
\title{Exploring Kernel Density Estimation}
\usage{
kdExplore(data, options = NULL)
}
\arguments{
\item{data}{A vector of numerical values from which to form the density plot.}

\item{options}{Options that will be passed to \code{shiny::shinyApp}.}
}
\value{
side effects
}
\description{
See how density plots are built from kernels.
}
\examples{
## small custom dataset:
## myData <- c(1, 3, 5, 6, 6.2, 7, 9)
## kdExplore(myData)

## random exponential data:
## kdExplore(rexp(50, rate = 0.2))
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
