% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{make_val_tree}
\alias{make_val_tree}
\title{Building Random Trees}
\usage{
make_val_tree(min_children = 1,
                     max_children = 4,
                     min_depth = 3,
                     values = 1:5,
                     fillcolor = "palegreen",
                     fontcolor = "black",
                     seed = NULL)
}
\arguments{
\item{min_children}{Minimum number of children when node will not be a leaf.}

\item{max_children}{Maximum number of children when node will not be a leaf.}

\item{min_depth}{Nodes at less than this specified depth will not be leaves.}

\item{values}{Numerical vector of possible values to associate with each node.}

\item{fillcolor}{Fill color for nodes.}

\item{fontcolor}{Fonot color for text showing node value and node name.}

\item{seed}{Option to set the random seed.}
}
\value{
A tree object of class "Node" and "R6"
}
\description{
Utility function to make small directed acyclic graphs, using
the data.tree package. Nodes are
provided with randomly-selected numerical values.
}
\examples{
\dontrun{
tr <- make_val_tree(
  values = 1:10,
  seed = 4040,
  fillcolor = "lightblue"
)
library(DiagrammeR)
plot(tr)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
