% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NamePhone}
\alias{NamePhone}
\title{Names and Phone Numbers}
\format{A data frame with 50 rows and 2 variables}
\description{
Sample data for regular expression practice.

\itemize{
      \item name.  Last name followed by first.
      \item phone. Phone number with area code, in several formats.
}
}
\keyword{datasets.}
