% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whales.R
\docType{class}
\name{Ocean}
\alias{Ocean}
\title{Whales in an Ocean}
\format{\code{\link{R6Class}} object.}
\usage{
Ocean
}
\value{
Object of \code{\link{R6Class}} with methods for simulation.
}
\description{
R6 Object for simulating a population of whales.
}
\section{Methods}{

\itemize{
\item{\code{new}: }{Instantiates an Ocean object.  Parameters are:}
\itemize{
\item{\code{dims}: }{A numeric vector of length 2 giving the length
and width
 of the ocean.}
\item{\code{males}: }{An integer giving the number of males (to be
created with
defaults) or a list of \code{\link{Male}} whale objects.}
\item{\code{females}: }{An integer giving the number of females (to be
created with
defaults) or a list of \code{\link{Female}} whale objects.}
\item{\code{starve}: }{A non-negative number, used to determine
probability that an individual starves in a given
generation. The larger the
value, the lower the carrying-capacity of the population
will be.}
}
}
}

\examples{
\dontrun{
library(ggplot2)
oceanSim <- function(
  steps = 100, males = 10,
  females = 10, starve = 5,
  animate = TRUE, seed = NULL
  ) {
 if ( !is.null(seed) ) {
   set.seed(seed)
  }
  ocean <- Ocean$new(dims = c(100, 100), males = males,
                     females = females, starve = starve)
  population <-numeric(steps)
  for ( i in 1:steps ) {
    population[i] <- ocean$malePop + ocean$femalePop
    if ( animate ) ocean$plot()
    if ( population[i] == 0 ) break
    ocean$advance()
    if ( animate ) Sys.sleep(0.5)
  }
  pop <- population[1:i]
  df <- data.frame(time = 1:length(pop),
                  pop)
  ggplot(df, aes(x = time, y = pop)) + geom_line() +
    labs(x = "Time", y = "Whale Population")
}
oceanSim()
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
\keyword{data}
