% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozLookup.R
\name{ozLookup}
\alias{ozLookup}
\title{Look Up Words in the Wizard of Oz}
\usage{
ozLookup(word)
}
\arguments{
\item{word}{a character vecotr of length 1, the word to be looked up}
}
\value{
side effects
}
\description{
Look up words in the Wizard of Oz.
This is primarily a utility-function for use of webR in the book.
}
\examples{
ozLookup(word = "humbug")
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
