% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqExplore.R
\name{qqExplore}
\alias{qqExplore}
\title{Exploring Quantile-Quantile Plots}
\usage{
qqExplore(data, options = NULL)
}
\arguments{
\item{data}{A vector of numerical values from which to form the qq-plot.}

\item{options}{Options that will be passed to \code{shiny::shinyApp}.}
}
\value{
side effects
}
\description{
Understand why qq-plots bend the way they do.
}
\examples{
\dontrun{
## bimodal data:
bimodal <- c(rnorm(50, 5, 1), rnorm(50, 10, 1))
kdExplore(bimodal)

## random exponential data:
qqExplore(rexp(300, rate = 0.2))
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
