% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{railtrail}
\alias{railtrail}
\title{Volume of Users of a Rail Trail}
\format{
A data frame with 90 observations on the following variables.
  \itemize{
    \item hightemp daily high temperature (in degrees Fahrenheit)
    \item lowtemp daily low temperature (in degrees Fahrenheit)
    \item avgtemp average of daily low and daily high temperature (in degrees Fahrenheit)
    \item season spring, summer or fall
    \item cloudcover measure of cloud cover (in oktas)
    \item precip measure of precipitation (in inches)
    \item volume estimated number of trail users that day (number of breaks recorded)
    \item weekday logical indicator of whether the day was a non-holiday weekday
    \item dayType one of "weekday" or "weekend"
  }
}
\source{
Pioneer Valley Planning Commission
}
\usage{
data(railtrail)
}
\description{
This data table is modifed slightly from mosaicData::RailTrail,
(see http://cran.r-project.org/web/packages/mosaicData/mosaicData.pdf).
Description below is drawn from the mosaicData help file.
}
\details{
The Pioneer Valley Planning Commission (PVPC) collected data north of Chestnut Street in
Florence, MA for ninety days from April 5, 2005 to November 15, 2005. Data collectors set up a
laser sensor, with breaks in the laser beam recording when a rail-trail user passed the
data collection station.


There is a potential for error when two users trigger the infrared beam at exactly the same time since the counter
would only logs one of the crossings.  The collectors left the motion detector out during the winter, but because the
counter drops data when the temperature falls below 14 degrees Fahrenheit, there is no data for the cold winter months.
}
\examples{
data(railtrail)

}
\references{
http://www.fvgreenway.org/pdfs/Northampton-Bikepath-Volume-Counts%20_05_LTA.pdf
}
