% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleSim.R
\name{triangleSim}
\alias{triangleSim}
\title{Chance of a Triangle}
\usage{
triangleSim(reps = 10000, table = FALSE, seed = NULL)
}
\arguments{
\item{reps}{number of simulations to perform}

\item{table}{Does the user want a table of the results?}

\item{seed}{The user may provide a seed-value for random-number
generation.}
}
\value{
side effects
}
\description{
Break a unit length at two random points:  what's the chance
that the three segments produced can form a triangle?
}
\examples{
triangleSim(seed = 3030)
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
