% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle_bounce.R
\name{turtle_bounce}
\alias{turtle_bounce}
\title{Bouncing Turtle (Turtle Graphics)}
\usage{
turtle_bounce(side = 60, step= 10)
}
\arguments{
\item{side}{side-lengths of the containing square}

\item{step}{length of one turtle step (side-length/2 must be
a multiple of step).}
}
\value{
side effects
}
\description{
A turtle walks randomly, but bounces back from the edge
of its containing field.
}
\examples{
## turtle_bounce(side = 80, step = 10)
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
