% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle_drunk.R
\name{turtle_drunk}
\alias{turtle_drunk}
\title{Drunken Turtle (Turtle Graphics)}
\usage{
turtle_drunk(side, step)
}
\arguments{
\item{side}{side-lengths of the containing square}

\item{step}{length of one turtle step}
}
\value{
side effects
}
\description{
A turtle takes steps of a fixed length, but turns at a
random angle after each step.
}
\examples{
\dontrun{
turtle_drunk(side = 100, step = 10)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
