% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whales.R
\docType{class}
\name{Whale}
\alias{Whale}
\alias{Female}
\alias{Male}
\alias{Male}
\alias{Female}
\title{Whales in an Ocean}
\format{\code{\link{R6Class}} object.}
\usage{
Whale

Male

Female
}
\value{
Object of \code{\link{R6Class}} with methods for simulation.
}
\description{
R6 Objects for modelling a whale.  Female and Male
inherit from whale.  Use Female and Male  if you want to
provide custom lists of male and female whales when you instantiate
an ocean.
}
\section{Methods}{

\itemize{
\item{\code{new}: }{Instantiates an Whale object.  Parameters are:}
\itemize{
\item{\code{position}: }{A numeric vector of length 2 giving the initial
position.  (Make sure that it's within the dimensions of the ocean.)}
\item{\code{age}: }{Initial age of the whale.}
\item{\code{lifespan}: }{Lifespan of the whale.}
\item{\code{range}: }{Distance at which a female can detect an
eligible male.}
\item{\code{maturity}: }{Age of whale at which reproduction is possible.}
\item{\code{stepSize}: }{Number of units the whale move sin each gneration.}
}
}
}

\examples{
\dontrun{
initialMales <- vector(mode = "list", length = 10)
ages <- c(rep(3, 5), c(rep(10, 5))
for (i in 1:10) {
  initialMales[[i]] <- Male$new(
    position = runif(2, min = 0, max = 100),
    age = ages[i],
    lifespan = 40,
    range = 12,
    maturity = 10,
    stepSize = 7
  )
}
initialFemales <- vector(mode = "list", length = 10)
for (i in 1:10) {
  initialFemales[[i]] <- Female$new(
    position = runif(2, min = 0, max = 100),
    age = ages[i],
    lifespan = 40,
    maturity = 10,
    range = 12,
    stepSize = 3
  )
}

library(ggplot2)
oceanSim <- function(
  steps = 100,
  males = initialMales,
  females = initialFemales,
  starve = 5,
  animate = TRUE,
  seed = NULL
  ) {
 if ( !is.null(seed) ) {
   set.seed(seed)
  }
  ocean <- Ocean$new(dims = c(100, 100), males = males,
                     females = females, starve = starve)
  population <-numeric(steps)
  for ( i in 1:steps ) {
    population[i] <- ocean$malePop + ocean$femalePop
    if ( animate ) ocean$plot()
    if ( population[i] == 0 ) break
    ocean$advance()
    if ( animate ) Sys.sleep(0.5)
  }
  pop <- population[1:i]
  df <- data.frame(time = 1:length(pop),
                  pop)
  ggplot(df, aes(x = time, y = pop)) + geom_line() +
    labs(x = "Time", y = "Whale Population")
}
oceanSim()
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
\keyword{data}
\keyword{datasets}
