% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcscr-package.R
\docType{package}
\name{bcscr-package}
\alias{bcscr-package}
\alias{bcscr}
\title{bcscr}
\description{
Functions and datasets to accompany the text Beginning Computer Science with R
}
\references{
\url{https://homerhanumat.github.io/r-notes}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/homerhanumat/bcscr}
}

}
\author{
\strong{Maintainer}: Homer White \email{homerhanumat@gmail.com}


Homer White (\email{homerhanumat@gmail.com})
}
\keyword{internal}
\keyword{package}
