% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean.R
\docType{class}
\name{Ocean}
\alias{Ocean}
\title{Whales in an Ocean}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for simulation.
}
\description{
R6 Object for simulating a population of whales.
}
\examples{
\dontrun{
library(ggplot2)
oceanSim <- function(
  steps = 100, males = 10,
  females = 10, starve = 5,
  animate = FALSE, seed = NULL
  ) {
 if ( !is.null(seed) ) {
   set.seed(seed)
  }
  ocean <- Ocean$new(dims = c(100, 100), males = males,
                     females = females, starve = starve)
  population <-numeric(steps)
  for ( i in 1:steps ) {
    population[i] <- ocean$malePop + ocean$femalePop
    if ( animate ) ocean$plot()
    if ( population[i] == 0 ) break
    ocean$advance()
    if ( animate ) {
        ocean$plot()
        Sys.sleep(0.5)
      }
  }
  pop <- population[1:i]
  df <- data.frame(time = 1:length(pop),
                  pop)
  ggplot(df, aes(x = time, y = pop)) + geom_line() +
    labs(x = "Time", y = "Whale Population")
}
oceanSim(seed = 5050)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
\keyword{data}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dimensions}}{A vector of length two giving the
dimensions of the ocean.}

\item{\code{males}}{A list of R6 objects of class Male
containing the current population of male whales.}

\item{\code{females}}{A list of R6 objects of class Female
containing the current population of female whales.}

\item{\code{malePop}}{Current number of males
in the population.}

\item{\code{femalePop}}{Current number of females
in the population.}

\item{\code{starveParameter}}{Helps determine probability
for each
whale to die by starvation in the current generation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ocean-distance}{\code{Ocean$distance()}}
\item \href{#method-Ocean-new}{\code{Ocean$new()}}
\item \href{#method-Ocean-starvationProbability}{\code{Ocean$starvationProbability()}}
\item \href{#method-Ocean-advance}{\code{Ocean$advance()}}
\item \href{#method-Ocean-plot}{\code{Ocean$plot()}}
\item \href{#method-Ocean-clone}{\code{Ocean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ocean-distance"></a>}}
\if{latex}{\out{\hypertarget{method-Ocean-distance}{}}}
\subsection{Method \code{distance()}}{
Compute distance between two whales.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$distance(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{one whale}

\item{\code{b}}{another whale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ocean-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ocean-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ocean object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$new(dims = c(100, 100), males = 10, females = 10, starve = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dims}}{numerical vector of length 2 giving
the dimensions of the ocean}

\item{\code{males}}{either a numeric vector of length 1
giving the initial number of male whales, or a
list of R6 objects of class Male inheriting from
Whale.  If the former, then the specified number
of whales of class Male are instantiated, with
certain default characteristics.  If the latter,
then the given whales are cloned.}

\item{\code{females}}{either a numeric vector of length 1
giving the initial number of female whales, or a
list of R6 objects of class Female inheriting from
Whale.  If the former, then the specified number
of whales of class Female are instantiated, with
certain default characteristics.  If the latter,
then the given whales are cloned.}

\item{\code{starve}}{the starvation parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Ocean` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ocean-starvationProbability"></a>}}
\if{latex}{\out{\hypertarget{method-Ocean-starvationProbability}{}}}
\subsection{Method \code{starvationProbability()}}{
Compute the starvation probability.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$starvationProbability(popDensity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popDensity}}{population density}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a number.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ocean-advance"></a>}}
\if{latex}{\out{\hypertarget{method-Ocean-advance}{}}}
\subsection{Method \code{advance()}}{
advance the `Ocean` one step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$advance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the `Ocean` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ocean-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Ocean-plot}{}}}
\subsection{Method \code{plot()}}{
plot the `Ocean`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
side effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ocean-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ocean-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
