% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean.R
\docType{class}
\name{Ocean}
\alias{Ocean}
\title{Whales in an Ocean}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}} with methods for simulation.
}
\description{
R6 Object for simulating a population of whales.
}
\section{Properties}{

\itemize{
\item{\code{dimensions}: }{A vector of length two giving the
dimensions of the ocean.}
\item{\code{males}: }{A list of R6 objects of class Male
containing the current population of male whales.}
\item{\code{females}: }{A list of R6 objects of class Female
containing the current population of female whales.}
\item{\code{malePop}: }{Current number of males in the population.}
\item{\code{femalePop}: }{Current number of females in the population.}
\item{\code{starveParameter}: }{Helps determine probability for each
whale to die by starvation in the current generation.}
\item{\code{distance}: }{Computes distance between any two whales.}
}
}

\section{Methods}{

\itemize{
\item{\code{new}: }{Instantiates an Ocean object.  Parameters are:}
\itemize{
\item{\code{dims}: }{A numeric vector of length 2 setting the length
and width
 of the ocean.}
\item{\code{males}: }{An integer giving the number of males (to be
created with
defaults) or a list of \code{\link{Male}} whale objects.}
\item{\code{females}: }{An integer giving the number of females (to be
created with
defaults) or a list of \code{\link{Female}} whale objects.}
\item{\code{starve}: }{A non-negative number, used to determine
probability that an individual starves in a given
generation. The larger the
value, the lower the carrying-capacity of the population
will be.}
}
\item{\code{advance}: }{Advances the simulation by one generation.
Takes no arguments.}
\item{\code{plot}: }{Plots the current population.  Takes no arguments.}
}
}

\examples{
\dontrun{
library(ggplot2)
oceanSim <- function(
  steps = 100, males = 10,
  females = 10, starve = 5,
  animate = FALSE, seed = NULL
  ) {
 if ( !is.null(seed) ) {
   set.seed(seed)
  }
  ocean <- Ocean$new(dims = c(100, 100), males = males,
                     females = females, starve = starve)
  population <-numeric(steps)
  for ( i in 1:steps ) {
    population[i] <- ocean$malePop + ocean$femalePop
    if ( animate ) ocean$plot()
    if ( population[i] == 0 ) break
    ocean$advance()
    if ( animate ) {
        ocean$plot()
        Sys.sleep(0.5)
      }
  }
  pop <- population[1:i]
  df <- data.frame(time = 1:length(pop),
                  pop)
  ggplot(df, aes(x = time, y = pop)) + geom_line() +
    labs(x = "Time", y = "Whale Population")
}
oceanSim(seed = 5050)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-distance}{\code{Ocean$distance()}}
\item \href{#method-new}{\code{Ocean$new()}}
\item \href{#method-starvationProbability}{\code{Ocean$starvationProbability()}}
\item \href{#method-advance}{\code{Ocean$advance()}}
\item \href{#method-plot}{\code{Ocean$plot()}}
\item \href{#method-clone}{\code{Ocean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distance"></a>}}
\subsection{Method \code{distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$distance(a, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$new(dims = c(100, 100), males = 10, females = 10, starve = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-starvationProbability"></a>}}
\subsection{Method \code{starvationProbability()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$starvationProbability(popDensity)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-advance"></a>}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$advance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ocean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
