% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whales.R
\docType{class}
\name{Whale}
\alias{Whale}
\alias{Female}
\alias{Male}
\title{Whales in an Ocean}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for simulation.

Boolean

sex of the resulting baby
}
\description{
R6 Objects for modelling a whale.  Female and Male
inherit from whale.  Use Female and Male  if you want to
provide custom lists of male and female whales when you instantiate
an ocean.
}
\examples{
\dontrun{
initialMales <- vector(mode = "list", length = 10)
ages <- c(rep(3, 5), c(rep(10, 5)))
for (i in 1:10) {
  initialMales[[i]] <- Male$new(
    position = runif(2, min = 0, max = 100),
    age = ages[i],
    lifespan = 40,
    range = 12,
    maturity = 10,
    stepSize = 7
  )
}
initialFemales <- vector(mode = "list", length = 10)
for (i in 1:10) {
  initialFemales[[i]] <- Female$new(
    position = runif(2, min = 0, max = 100),
    age = ages[i],
    lifespan = 40,
    maturity = 10,
    range = 12,
    stepSize = 3
  )
}

library(ggplot2)
oceanSim <- function(
  steps = 100,
  males = 10,
  females = 10,
  starve = 5,
  animate = FALSE,
  seed = NULL
  ) {
 if ( !is.null(seed) ) {
   set.seed(seed)
  }
  if (is.list(males)) {
    processed_males <- vector(mode = "list", length = length(males))
    for (i in seq_along(males)) {
      processed_males[[i]] <- males[[i]]$clone()
    }
  } else {
    processed_males <- males
  }
  if (is.list(females)) {
    processed_females <- vector(mode = "list", length = length(females))
    for (i in seq_along(females)) {
      processed_females[[i]] <- females[[i]]$clone()
    }
  } else {
    processed_females <- females
  }

  ocean <- Ocean$new(
    dims = c(100, 100),
    males = processed_males,
    females = processed_females,
    starve = starve
   )
  population <-numeric(steps)
  for (i in 1:steps) {
    population[i] <- ocean$malePop + ocean$femalePop
    if (animate) ocean$plot()
    if (population[i] == 0) break
    ocean$advance()
    if (animate) {
      ocean$plot()
      Sys.sleep(0.5)
     }
  }
  pop <- population[1:i]
  df <- data.frame(
    time = 1:length(pop),
    pop
   )
  ggplot(df, aes(x = time, y = pop)) +
    geom_line() +
    labs(x = "Time", y = "Whale Population")
}
oceanSim(males = initialMales, females = initialFemales, seed = 5050)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
\keyword{data}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position}}{A numeric vector of length 2 giving the
 initial
position.  (Make sure that it's within the
dimensions of the ocean.)}

\item{\code{age}}{initial age of the whale}

\item{\code{lifespan}}{lifespan of the whale}

\item{\code{range}}{Distance at which a female can detect an
eligible male.}

\item{\code{maturity}}{Age of whale at which reproduction
is possible.}

\item{\code{stepSize}}{Number of units the whale moves
in each generation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Whale-new}{\code{Whale$new()}}
\item \href{#method-Whale-move}{\code{Whale$move()}}
\item \href{#method-Whale-clone}{\code{Whale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whale-new"></a>}}
\if{latex}{\out{\hypertarget{method-Whale-new}{}}}
\subsection{Method \code{new()}}{
initialize a `Whale`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whale$new(
  position = NA,
  age = 3,
  lifespan = 40,
  range = 5,
  maturity = 10,
  stepSize = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{position of whale}

\item{\code{age}}{age of whale}

\item{\code{lifespan}}{lifespan of whale}

\item{\code{range}}{range of whale}

\item{\code{maturity}}{how old whale must be in order to mate}

\item{\code{stepSize}}{how far whale moves in one step}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new `Whale` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whale-move"></a>}}
\if{latex}{\out{\hypertarget{method-Whale-move}{}}}
\subsection{Method \code{move()}}{
move a whale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whale$move(dims, r = self$stepSize)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dims}}{dimensions of the ocean}

\item{\code{r}}{how far the whale wants to move}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whale-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Whale-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[bcscr:Whale]{bcscr::Whale}} -> \code{Male}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sex}}{sex of the whale}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Male-clone}{\code{Male$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="initialize"><a href='../../bcscr/html/Whale.html#method-Whale-initialize'><code>bcscr::Whale$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="move"><a href='../../bcscr/html/Whale.html#method-Whale-move'><code>bcscr::Whale$move()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Male-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Male-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Male$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[bcscr:Whale]{bcscr::Whale}} -> \code{Female}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sex}}{sex of the whale}

\item{\code{timeToFertility}}{how long until whale can breed}

\item{\code{infertilityPeriod}}{number of steps until a female
can breed again after having borne a child}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Female-maleNear}{\code{Female$maleNear()}}
\item \href{#method-Female-mate}{\code{Female$mate()}}
\item \href{#method-Female-clone}{\code{Female$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="initialize"><a href='../../bcscr/html/Whale.html#method-Whale-initialize'><code>bcscr::Whale$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="move"><a href='../../bcscr/html/Whale.html#method-Whale-move'><code>bcscr::Whale$move()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Female-maleNear"></a>}}
\if{latex}{\out{\hypertarget{method-Female-maleNear}{}}}
\subsection{Method \code{maleNear()}}{
determine whether a male is near
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Female$maleNear(males, dist)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{males}}{the males in the ocean}

\item{\code{dist}}{the distance-fidning function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Female-mate"></a>}}
\if{latex}{\out{\hypertarget{method-Female-mate}{}}}
\subsection{Method \code{mate()}}{
make female whale mate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Female$mate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Female-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Female-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Female$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
