% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whales.R
\docType{class}
\name{Whale}
\alias{Whale}
\alias{Female}
\alias{Male}
\title{Whales in an Ocean}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}} with methods for simulation.
}
\description{
R6 Objects for modelling a whale.  Female and Male
inherit from whale.  Use Female and Male  if you want to
provide custom lists of male and female whales when you instantiate
an ocean.
}
\section{Methods}{

\itemize{
\item{\code{new}: }{Instantiates an Whale object.  Parameters are:}
\itemize{
\item{\code{position}: }{A numeric vector of length 2 giving the initial
position.  (Make sure that it's within the dimensions of the ocean.)}
\item{\code{age}: }{Initial age of the whale.}
\item{\code{lifespan}: }{Lifespan of the whale.}
\item{\code{range}: }{Distance at which a female can detect an
eligible male.}
\item{\code{maturity}: }{Age of whale at which reproduction is possible.}
\item{\code{stepSize}: }{Number of units the whale move sin each gneration.}
}
}
}

\examples{
\dontrun{
initialMales <- vector(mode = "list", length = 10)
ages <- c(rep(3, 5), c(rep(10, 5)))
for (i in 1:10) {
  initialMales[[i]] <- Male$new(
    position = runif(2, min = 0, max = 100),
    age = ages[i],
    lifespan = 40,
    range = 12,
    maturity = 10,
    stepSize = 7
  )
}
initialFemales <- vector(mode = "list", length = 10)
for (i in 1:10) {
  initialFemales[[i]] <- Female$new(
    position = runif(2, min = 0, max = 100),
    age = ages[i],
    lifespan = 40,
    maturity = 10,
    range = 12,
    stepSize = 3
  )
}

library(ggplot2)
oceanSim <- function(
  steps = 100,
  males = 10,
  females = 10,
  starve = 5,
  animate = FALSE,
  seed = NULL
  ) {
 if ( !is.null(seed) ) {
   set.seed(seed)
  }
  ocean <- Ocean$new(dims = c(100, 100), males = males,
                     females = females, starve = starve)
  population <-numeric(steps)
  for ( i in 1:steps ) {
    population[i] <- ocean$malePop + ocean$femalePop
    if ( animate ) ocean$plot()
    if ( population[i] == 0 ) break
    ocean$advance()
    if ( animate ) {
        ocean$plot()
        Sys.sleep(0.5)
      }
  }
  pop <- population[1:i]
  df <- data.frame(time = 1:length(pop),
                  pop)
  ggplot(df, aes(x = time, y = pop)) + geom_line() +
    labs(x = "Time", y = "Whale Population")
}
oceanSim(males = initialMales, females = initialFemales, seed = 5050)
}
}
\author{
Homer White \email{homerhanumat@gmail.com}
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Whale$new()}}
\item \href{#method-move}{\code{Whale$move()}}
\item \href{#method-clone}{\code{Whale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whale$new(
  position = NA,
  age = 3,
  lifespan = 40,
  range = 5,
  maturity = 10,
  stepSize = 5
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move"></a>}}
\subsection{Method \code{move()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whale$move(dims, r = self$stepSize)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[bcscr:Whale]{bcscr::Whale}} -> \code{Male}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{Male$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="initialize">}\href{../../bcscr/html/Whale.html#method-initialize}{\code{bcscr::Whale$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="move">}\href{../../bcscr/html/Whale.html#method-move}{\code{bcscr::Whale$move()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Male$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[bcscr:Whale]{bcscr::Whale}} -> \code{Female}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-maleNear}{\code{Female$maleNear()}}
\item \href{#method-mate}{\code{Female$mate()}}
\item \href{#method-clone}{\code{Female$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="initialize">}\href{../../bcscr/html/Whale.html#method-initialize}{\code{bcscr::Whale$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bcscr" data-topic="Whale" data-id="move">}\href{../../bcscr/html/Whale.html#method-move}{\code{bcscr::Whale$move()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-maleNear"></a>}}
\subsection{Method \code{maleNear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Female$maleNear(males, dist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mate"></a>}}
\subsection{Method \code{mate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Female$mate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Female$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
